+++ src/dxp/GameRequest.h
@@ -1,5 +1,4 @@
 #pragma once
-#include <cstddef>
 #include <memory>       // std::unique_ptr
 #include <string>
 #include "MessageInterface.h"
@@ -13,11 +12,11 @@ public:
         // views
         const std::string& name_initiator() const;
         char color_follower() const;
-        size_t minutes() const;
-        size_t moves() const;
+        int minutes() const;
+        int moves() const;
         bool setup() const;
         const std::string& position() const;
-        static std::string str(const std::string&, char, size_t, size_t, bool, const std::string&);
+        static std::string str(const std::string&, char, int, int, bool, const std::string&);
 
 private:
         // private constructor to enforce factory creation
@@ -26,11 +25,11 @@ private:
         // implementation
         virtual std::string header() const;
         virtual std::string body() const;
-        static std::string body(const std::string&, char, size_t, size_t, bool, const std::string&);
+        static std::string body(const std::string&, char, int, int, bool, const std::string&);
 
         static std::unique_ptr<MessageInterface> create(const std::string&);
         static const std::string HEADER;
-        static const size_t PROTOCOL_VERSION = 1;
+        static const int PROTOCOL_VERSION = 1;
         static const bool REGISTERED;
         static const char INITIAL;
         static const char SPECIAL;
@@ -42,8 +41,8 @@ private:
         // representation
         std::string name_initiator_;
         char color_follower_;
-        size_t minutes_;
-        size_t moves_;
+        int minutes_;
+        int moves_;
         bool setup_;
         std::string position_;
 };
