+++ src/generate/PawnJumps.h
@@ -19,7 +19,7 @@ class Driver<Color, Material::PAWN, Move::JUMPS, Rules, Board>
 public:
         static void generate(const Position<Board>&, Stack&);
         static void generate(const Position<Board>&, capture::State<Rules, Board>&, Stack&);
-        static size_t count(const Position<Board>&);
+        static int count(const Position<Board>&);
         static bool detect(const Position<Board>&);
 
 private:
@@ -28,7 +28,7 @@ private:
         // tag dispatching based on whether men can capture kings
         static void generate_targets(const Position<Board>&, capture::State<Rules, Board>&, Stack&);
         static void generate_targets(const Position<Board>&, capture::State<Rules, Board>&, Stack&, Int2Type<false>);
-        static void generate_targets(const Position<Board>&, capture::State<Rules, Board>&, Stack&, Int2Type<true>);
+        static void generate_targets(const Position<Board>&, capture::State<Rules, Board>&, Stack&, Int2Type<true >);
 
         // tag dispatching based on man capture directions
         static void generate_dirs(BitBoard, capture::State<Rules, Board>&, Stack&);
@@ -38,24 +38,24 @@ private:
         static void generate_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::DIRS_UP  >);
         static void generate_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::DIRS_DOWN>);
 
-        template<size_t> static void generate_dir(BitBoard, capture::State<Rules, Board>&, Stack&);
-        template<size_t> static void generate_next(BitBoard, capture::State<Rules, Board>&, Stack&);
+        template<int> static void generate_dir(BitBoard, capture::State<Rules, Board>&, Stack&);
+        template<int> static void generate_next(BitBoard, capture::State<Rules, Board>&, Stack&);
 
         // tag dispatching based on promotion condition
-        template<size_t> static bool scan_next(BitBoard, capture::State<Rules, Board>&, Stack&);
-        template<size_t> static bool scan_next(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::PROMOTE_BR>);
-        template<size_t> static bool scan_next(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::PROMOTE_EP>);
+        template<int> static bool scan_next(BitBoard, capture::State<Rules, Board>&, Stack&);
+        template<int> static bool scan_next(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::PROMOTE_BR>);
+        template<int> static bool scan_next(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::PROMOTE_EP>);
 
         // tag dispatching based on man scan directions
-        template<size_t> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&);
-        template<size_t> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_ALL >);
-        template<size_t> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_REST>);
-        template<size_t> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_SIDE>);
-        template<size_t> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_UP  >);
-        template<size_t> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_DOWN>);
+        template<int> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&);
+        template<int> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_ALL >);
+        template<int> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_REST>);
+        template<int> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_SIDE>);
+        template<int> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_UP  >);
+        template<int> static bool scan_dirs(BitBoard, capture::State<Rules, Board>&, Stack&, Int2Type<rules::SCAN_DOWN>);
 
-        template<size_t> static bool scan_dir(BitBoard, capture::State<Rules, Board>&, Stack&);
-        template<size_t> static bool scan(BitBoard, capture::State<Rules, Board>&, Stack&);
+        template<int> static bool scan_dir(BitBoard, capture::State<Rules, Board>&, Stack&);
+        template<int> static bool scan(BitBoard, capture::State<Rules, Board>&, Stack&);
 
         // tag dispatching based on whether men can capture kings
         static bool detect_targets(const Position<Board>&);
@@ -70,7 +70,7 @@ private:
         static bool detect_dirs(BitBoard, BitBoard, BitBoard, Int2Type<rules::DIRS_UP  >);
         static bool detect_dirs(BitBoard, BitBoard, BitBoard, Int2Type<rules::DIRS_DOWN>);
 
-        template<size_t> static bool detect_dir(BitBoard, BitBoard, BitBoard);
+        template<int> static bool detect_dir(BitBoard, BitBoard, BitBoard);
 };
 
 }       // namespace generate
